Fixing some implicit casts with data loss.
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_softfloat@1567 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/dimension/i860dec.cpp
@@ -128,7 +128,7 @@ UINT32 i860_cpu_device::ifetch_notrap(const UINT32 pc) {
 }
 
 UINT32 i860_cpu_device::ifetch(const UINT32 pc) {
-    return pc & 4 ? ifetch64(pc) >> 32 : ifetch64(pc);
+    return (UINT32)(pc & 4 ? ifetch64(pc) >> 32 : ifetch64(pc));
 }
 
 UINT64 i860_cpu_device::ifetch64(const UINT32 pc, const UINT32 vaddr, int const cidx) {
